First attempt at creating a starting point for
running scripts for any given project by using YAML config files.
The idea is that all of the parameters that could change from
run to run can be loaded from a yaml file, increasing reproducibility.
Also automatically sets up the sql engine based on the given settings
and sets up the hpc jobs based on settings in the yaml file.